requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
openai==1.3.0
jieba==0.42.1
textblob==0.15.3
emoji==1.7.0
python-dotenv==0.19.0